Draw a horizontal circle

A utility function for drawing a horizontal circle in a 3D graph

circle3d(center, radius, segments = 100, fill = FALSE, ...)

Arguments

center
A vector of length 3.

radius
A positive number.

segments
An integer specifying the number of line segments to use to draw the circle (default, 100).

fill
logical; if TRUE, the circle is filled (the default is FALSE).

...
rgl material properties for the circle.