Data on Coffee, Stress and Heart Damage

A dataset, used for examples in Friendly, Fox & Monette (2013), giving an index of cardiac damage (Heart) in relation to a measure of daily coffee consumption (Coffee) and Stress, a measure of perceived occupational stress, in a contrived sample of \(n = 20\) university people.

Usage

data("coffee")

Format

A data frame with 20 observations on the following 4 variables.

Group

university group, a factor with levels Grad_Student Professor Student

Coffee

a measure of daily coffee consumption

Stress

a measure of perceived occupational stress

Heart

an index of cardiac damage

Details

The main goal for analysis of this teaching example would be to determine whether or not coffee is good or bad for your heart, and stress represents one potential confounding variable among others (age, smoking, etc.) that might be useful to control statistically.

Friendly et al. (2013) use this data to illustrate (a) data ellipses in data space and the corresponding confidence ellipses in parameter (\(\beta\)) space; (b) effects of measurement error in a predictor or response; (c) added-variable plots and more.

Source

This dataset was constructed by Georges Monette, and was modified from that in his spida2 package, https://github.com/gmonette/spida2.

References

Friendly, M., Monette, G., & Fox, J. (2013). Elliptical Insights: Understanding Statistical Methods Through Elliptical Geometry. Statistical Science, 28(1), 1–39. https://doi.org/10.1214/12-STS402

Examples

data(coffee)
## maybe str(coffee) ; plot(coffee) ...