Draw a 3D cone

Draws a cone in 3D from a base point to a tip point, with a given radius at the base. This is used to draw nice arrow heads in arrows3d.

cone3d(base, tip, radius = 10, col = "grey", scale = NULL, ...)

Arguments

base

coordinates of base of the cone

tip

coordinates of tip of the cone

radius

radius of the base

col

color

scale

scale factor for base and tip

...

rgl arguments passed down; see rgl.material

Value

returns the integer object ID of the shape that was added to the scene

See also

Author

January Weiner, borrowed from from the pca3d package

Examples

# none yet