Gram-Schmidt Orthogonalization of a Matrix

Calculates a matrix with uncorrelated columns using the Gram-Schmidt process

gsorth(y, order, recenter = TRUE, rescale = TRUE, adjnames = TRUE)

Arguments

y
a numeric matrix or data frame

order
if specified, a permutation of the column indices of y

recenter
logical; if TRUE, the result has same means as the original y, else means = 0 for cols 2:p

rescale
logical; if TRUE, the result has same sd as original, else, sd = residual sd

adjnames
logical; if TRUE, colnames are adjusted to Y1, Y2.1, Y3.12, ...

Value

a matrix/data frame with uncorrelated columns

Details

This function, originally from the heplots package has now been deprecated in matlib. Use GramSchmidt instead.

Examples

# NOT RUN {
 set.seed(1234)
 A <- matrix(c(1:60 + rnorm(60)), 20, 3)
 cor(A)
 G <- gsorth(A)
 zapsmall(cor(G))

# }