Create a Symbolic Matrix for LaTeX

Constructs the LaTeX code for a symbolic matrix, like:

 \begin{pmatrix}
   x_{11}  & x_{12}  & \dots  & x_{1m}  \
   x_{21}  & x_{22}  & \dots  & x_{2m}  \
   \vdots & \vdots & \ddots & \vdots \
   x_{n1}  & x_{n2}  & \dots  & x_{nm}
 \end{pmatrix}
 

When rendered in LaTeX, this produces: $$ \begin{pmatrix} x_{11} & x_{12} & \cdots & x_{1m} \\ x_{21} & x_{22} & \cdots & x_{2m} \\ \vdots & \vdots & & \vdots \\ x_{n1} & x_{n2} & \cdots & x_{nm} \\ \end{pmatrix} $$

Alternatively, the number of rows and/or columns can be integers, generating a matrix of given size.

The function prints the resulting code to the console, or, using write_clip, can be copied to the console, or can be used used in a markdown chunk in a Rmd or qmd document, e.g,

```{r results = "asis"}
symbolicMatrix("\lambda", nrow=2, ncol=2,
               diag=TRUE,
               lhs = "\boldsymbol{\Lambda}")
```

This generates $$ \boldsymbol{\Lambda} = \begin{pmatrix} \lambda_{1} & 0 \\ 0 & \lambda_{2} \\ \end{pmatrix} $$

Usage

symbolicMatrix(
  symbol = "x",
  nrow = "n",
  ncol = "m",
  matrix = "pmatrix",
  diag = FALSE,
  comma = FALSE,
  exponent,
  transpose = FALSE,
  lhs,
  print = TRUE
)

Arguments

symbol

name for matrix elements, character string. For LaTeX symbols, the backslash must be doubled because it is an escape character in R. That is, you must use symbol = "\\beta" to get \(\beta\).

nrow

Number of rows, a single character representing rows symbolically, or an integer, generating that many rows.

ncol

Number of columns, a single character representing columns symbolically, or an integer, generating that many columns.

matrix

Character string giving the LaTeX matrix environment used in \begin{}, \end{}. Typically one of:

"pmatrix"

uses parentheses: "(", ")"

"bmatrix"

uses square brackets: "[", "]"

"Bmatrix"

uses braces: "{", "}"

"vmatrix"

uses vertical bars: "|", "|"

"Vmatrix"

uses double vertical bars: "||", "||"

"matrix"

generates a plain matrix without delimeters

diag

logical; if TRUE, off-diagonal elements are all 0 (and nrow must == ncol)

comma

logical; if TRUE, commas are inserted between row and column subscripts

exponent

if specified, e.g., "-1", or "1/2", the exponent is applied to the matrix

transpose

if TRUE, the transpose symbol "\top" is appended to the matrix; may also be a character string, e.g., "T", "\prime", "\textsf{T}"

lhs

character; an optional LaTeX expression, e.g, "\boldsymbol{\Lamda}", for left-hand side of an equation with the generated matrix on the right-hand side.

print

logical; print the LaTeX code for the matrix on the console?; default: TRUE

Value

Returns invisibly the LaTeX representation of the matrix as a character string. If you assign to a variable, you can use write_clip to copy it to the clipboard.

As a side-effect, by default (unless print = FALSE) the function uses cat() to display the result at the console.

Details

This implementation assumes that the LaTeX amsmath package will be available because it uses the shorthands \begin{pmatrix}, ... rather than

\left(
  \begin{array}
  ...
  \end{array}
\right)

.

This function is experimental. Other features may be added. E.g., it would be nice to:

Specify exponents for the matrix elements, e.g, a diagonal matrix of square roots of eigenvalues, \lambda_i^{1/2} giving \(\lambda_i^{1/2}\)

See also

Author

John Fox

Examples

symbolicMatrix()
#> Error in symbolicMatrix(): could not find function "symbolicMatrix"

# return value
mat <- symbolicMatrix(print = FALSE)
#> Error in symbolicMatrix(print = FALSE): could not find function "symbolicMatrix"
str(mat)
#> Error in eval(expr, envir, enclos): object 'mat' not found
cat(mat)
#> Error in eval(expr, envir, enclos): object 'mat' not found
# copy to clipboard
#clipr::write_clip(mat)    # can't be done in non-interactive mode

# can use a complex symbol
symbolicMatrix("\\widehat{\\beta}", 2, 4)
#> Error in symbolicMatrix("\\widehat{\\beta}", 2, 4): could not find function "symbolicMatrix"

# numeric rows/cols
symbolicMatrix(ncol=3)
#> Error in symbolicMatrix(ncol = 3): could not find function "symbolicMatrix"
symbolicMatrix(nrow=4)
#> Error in symbolicMatrix(nrow = 4): could not find function "symbolicMatrix"
symbolicMatrix(nrow=4, ncol=4)
#> Error in symbolicMatrix(nrow = 4, ncol = 4): could not find function "symbolicMatrix"

# diagonal matrices
symbolicMatrix(nrow=3, ncol=3, diag=TRUE)
#> Error in symbolicMatrix(nrow = 3, ncol = 3, diag = TRUE): could not find function "symbolicMatrix"
symbolicMatrix(nrow="n", ncol="n", diag=TRUE)
#> Error in symbolicMatrix(nrow = "n", ncol = "n", diag = TRUE): could not find function "symbolicMatrix"

# commas, exponents, transpose
symbolicMatrix("\\beta", comma=TRUE, exponent="-1")
#> Error in symbolicMatrix("\\beta", comma = TRUE, exponent = "-1"): could not find function "symbolicMatrix"
symbolicMatrix("\\beta", comma=TRUE, transpose=TRUE)
#> Error in symbolicMatrix("\\beta", comma = TRUE, transpose = TRUE): could not find function "symbolicMatrix"
symbolicMatrix("\\beta", comma=TRUE, exponent="-1", transpose=TRUE)
#> Error in symbolicMatrix("\\beta", comma = TRUE, exponent = "-1", transpose = TRUE): could not find function "symbolicMatrix"

# represent the SVD, X = U D V'  symbolically
X <- symbolicMatrix("x", "n", "p")
#> Error in symbolicMatrix("x", "n", "p"): could not find function "symbolicMatrix"
U <- symbolicMatrix("u", "n", "k")
#> Error in symbolicMatrix("u", "n", "k"): could not find function "symbolicMatrix"
D <- symbolicMatrix("\\lambda", "k", "k", diag=TRUE)
#> Error in symbolicMatrix("\\lambda", "k", "k", diag = TRUE): could not find function "symbolicMatrix"
V <- symbolicMatrix("v", "k", "p", transpose = TRUE)
#> Error in symbolicMatrix("v", "k", "p", transpose = TRUE): could not find function "symbolicMatrix"
cat("\\mathrm{SVD:}\n", X, "=\n", U, D, V)
#> Error in eval(expr, envir, enclos): object 'X' not found

# specify left hand side
symbolicMatrix("\\lambda", 3, 3, diag=TRUE, lhs = "\\boldsymbol{\\Lambda}")
#> Error in symbolicMatrix("\\lambda", 3, 3, diag = TRUE, lhs = "\\boldsymbol{\\Lambda}"): could not find function "symbolicMatrix"