Create a LaTeX Equation Wrapper

This function is designed to produce LaTeX expressions that can be copied/pasted into documents or used directly in .Rmd or .qmd documents to compile to equations. It wraps the equations in either a \begin{equation} ...\end{equation} or \begin{align} ...\end{align} environment.

In a code chunk, use the chunk options results='asis', echo=FALSE.

Usage

Eqn(
  ...,
  number = TRUE,
  label = NULL,
  align = FALSE,
  html_output = knitr::is_html_output(),
  mat_args = list()
)

Eqn_newline()

Arguments

...

comma separated expressions that are either a) a character vector, which will be automatically wrapped the expression inside a call to cat, b) a matrix object containing character or numeric information, which will be passed latexMatrix, or c) an object that was explicitly created via latexMatrix, which provides greater specificity.

Note that user defined functions that use cat within their body should return an empty character vector to avoid printing the returned object

number

logical; include equation number? Default: TRUE

label

character vector specifying the LaTeX label to use (e.g., eqn:myeqn), which can be reference via \ref{eqn:myeqn}. For compiled documents, if an HTML output is detected (see html_output) then the equations will be labelled via (\#eqn:myeqn) and references via \@ref(eq:binom)

align

logical; use the align environment with explicit &. Default: FALSE

html_output

logical; use labels for HTML outputs instead of the LaTeX? Automatically changed for compiled documents that support knitr

mat_args

list of arguments to be passed to latexMatrix to change the properties of the matrix input objects. Note that these inputs are used globally, and apply to each matrix objects supplied. If further specificity is required create latexMatrix objects directly

See also

Author

Phil Chalmers

Examples

# character input
Eqn('e=mc^2')
#> 
#> \begin{equation}
#> e=mc^2
#> \end{equation}

# Equation numbers & labels
Eqn('e=mc^2', number = FALSE)
#> 
#> \begin{equation*}
#> e=mc^2
#> \end{equation*}
Eqn('e=mc^2', label = 'eqn:einstein')
#> 
#> \begin{equation}
#> \label{eqn:einstein}
#> e=mc^2
#> \end{equation}
Eqn("X=U \\lambda V", label='eqn:svd')
#> 
#> \begin{equation}
#> \label{eqn:svd}
#> X=U \lambda V
#> \end{equation}

# html output (auto detected for documents)
Eqn('e=mc^2', label = 'eqn:einstein', html_output = TRUE)
#> 
#> \begin{equation}
#> (\#eqn:einstein)
#> e=mc^2
#> \end{equation}

# Multiple expressions
Eqn("e=mc^2",
    Eqn_newline(),
    "X=U \\lambda V", label='eqn:svd')
#> 
#> \begin{equation}
#> \label{eqn:svd}
#> e=mc^2 \\ 
#> X=U \lambda V
#> \end{equation}

# expressions that use cat() within their calls
Eqn(latexMatrix("u", "n", "k", lhs = 'SVD'),
    latexMatrix("\\lambda", "k", "k", diag=TRUE),
    latexMatrix("v", "k", "p", transpose = TRUE),
    label='eqn:svd')
#> 
#> \begin{equation}
#> \label{eqn:svd}
#> SVD = 
#> \begin{pmatrix} 
#>   u_{11} & u_{12} & \cdots & u_{1k} \\ 
#>   u_{21} & u_{22} & \cdots & u_{2k} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   u_{n1} & u_{n2} & \cdots & u_{nk} \\ 
#> \end{pmatrix}\begin{pmatrix} 
#>   \lambda_{1} & 0           & \cdots      & 0           \\ 
#>   0           & \lambda_{2} & \cdots      & 0           \\ 
#>   \vdots      & \vdots      & \ddots      & \vdots      \\ 
#>   0           & 0           & \cdots      & \lambda_{k} \\ 
#> \end{pmatrix}\begin{pmatrix} 
#>   v_{11} & v_{12} & \cdots & v_{1p} \\ 
#>   v_{21} & v_{22} & \cdots & v_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   v_{k1} & v_{k2} & \cdots & v_{kp} \\ 
#> \end{pmatrix}^\top
#> \end{equation}

# align equations using & operator
Eqn("X &= U \\lambda V",
    Eqn_newline(),
    latexMatrix("u", "n", "k", lhs = '&'),
    latexMatrix("\\lambda", "k", "k", diag=TRUE),
    latexMatrix("v", "k", "p", transpose = TRUE),
    align=TRUE)
#> 
#> \begin{align}
#> X &= U \lambda V \\ 
#> & = 
#> \begin{pmatrix} 
#>   u_{11} & u_{12} & \cdots & u_{1k} \\ 
#>   u_{21} & u_{22} & \cdots & u_{2k} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   u_{n1} & u_{n2} & \cdots & u_{nk} \\ 
#> \end{pmatrix}\begin{pmatrix} 
#>   \lambda_{1} & 0           & \cdots      & 0           \\ 
#>   0           & \lambda_{2} & \cdots      & 0           \\ 
#>   \vdots      & \vdots      & \ddots      & \vdots      \\ 
#>   0           & 0           & \cdots      & \lambda_{k} \\ 
#> \end{pmatrix}\begin{pmatrix} 
#>   v_{11} & v_{12} & \cdots & v_{1p} \\ 
#>   v_{21} & v_{22} & \cdots & v_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   v_{k1} & v_{k2} & \cdots & v_{kp} \\ 
#> \end{pmatrix}^\top
#> \end{align}

#  numeric/character matrix example
A <- matrix(c(2, 1, -1,
              -3, -1, 2,
              -2,  1, 2), 3, 3, byrow=TRUE)
b <- c(8, -11, -3)

cbind(A,b) |> Eqn()
#> 
#> \begin{equation}
#> \begin{pmatrix} 
#>   2 &   1 &  -1 &   8 \\ 
#>  -3 &  -1 &   2 & -11 \\ 
#>  -2 &   1 &   2 &  -3 \\ 
#> \end{pmatrix}
#> \end{equation}
cbind(A,'\\bigm|', b) |> Eqn()
#> 
#> \begin{equation}
#> \begin{pmatrix} 
#> 2  & 1  & -1 & \bigm| & 8   \\ 
#> -3 & -1 & 2  & \bigm| & -11 \\ 
#> -2 & 1  & 2  & \bigm| & -3  \\ 
#> \end{pmatrix}
#> \end{equation}

# change matrix brackets globally
cbind(A,b) |> Eqn(mat_args=list(matrix='bmatrix'))
#> 
#> \begin{equation}
#> \begin{bmatrix} 
#>   2 &   1 &  -1 &   8 \\ 
#>  -3 &  -1 &   2 & -11 \\ 
#>  -2 &   1 &   2 &  -3 \\ 
#> \end{bmatrix}
#> \end{equation}

# with showEqn()
showEqn(A, b, latex=TRUE) |> Eqn()
#> 
#> \begin{equation}
#> \begin{array}{lllllll}
#>  2 \cdot x_1 &+& 1 \cdot x_2 &-& 1 \cdot x_3  &=&    8 \\ 
#> -3 \cdot x_1 &-& 1 \cdot x_2 &+& 2 \cdot x_3  &=&  -11 \\ 
#> -2 \cdot x_1 &+& 1 \cdot x_2 &+& 2 \cdot x_3  &=&   -3 \\ 
#> \end{array}
#> \end{equation}