Draw a horizontal circle

A utility function for drawing a horizontal circle in the (x,y) plane in a 3D graph

Usage

circle3d(center, radius, segments = 100, fill = FALSE, ...)

Arguments

center

A vector of length 3.

radius

A positive number.

segments

An integer specifying the number of line segments to use to draw the circle (default, 100).

fill

logical; if TRUE, the circle is filled (the default is FALSE).

...

rgl material properties for the circle.

See also

Examples

ctr=c(0,0,0)
circle3d(ctr, 3, fill = TRUE)
circle3d(ctr - c(-1,-1,0), 3, col="blue")
circle3d(ctr + c(1,1,0),   3, col="red")

3D plot