Draw a corner showing the angle between two vectors

A utility function for drawing vector diagrams. Draws two line segments to indicate the angle between two vectors, typically used for indicating orthogonal vectors are at right angles in 2D and 3D diagrams.

corner(p1, p2, p3, d = 0.1, absolute = TRUE, ...)

Arguments

p1
Starting point of first vector

p2
End point of first vector, and also start of second vector

p3
End point of second vector

d
The distance from p2 along each vector for drawing their corner

absolute
logical; if TRUE, d is taken as an absolute distance along the vectors; otherwise it is calculated as a relative distance, i.e., a fraction of the length of the vectors. See pointOnLine for the precise definition.

...
Arguments passed to link[graphics]{lines} or to link[rgl]{lines3d}

Value

none

Details

In this implementation, the two vectors are specified by three points, p1, p2, p3, meaning a line from p1 to p2, and another line from p2 to p3.

See also

Other vector diagrams: Proj, arc, arrows3d, plot.regvec3d, pointOnLine, regvec3d, vectors3d, vectors

Examples

# none yet