Create a Symbolic Matrix for LaTeX

Constructs the LaTeX code for a symbolic matrix, whose elements are a symbol, with row and column subscripts. For example:

 \begin{pmatrix}
   x_{11}  & x_{12}  & \dots  & x_{1m}  \
   x_{21}  & x_{22}  & \dots  & x_{2m}  \
   \vdots & \vdots & \ddots & \vdots \
   x_{n1}  & x_{n2}  & \dots  & x_{nm}
 \end{pmatrix}
 

When rendered in LaTeX, this produces: $$ \begin{pmatrix} x_{11} & x_{12} & \cdots & x_{1m} \\ x_{21} & x_{22} & \cdots & x_{2m} \\ \vdots & \vdots & & \vdots \\ x_{n1} & x_{n2} & \cdots & x_{nm} \\ \end{pmatrix} $$

Alternatively, instead of characters, the number of rows and/or columns can be integers, generating a matrix of given size.

As well, instead of a character for the matrix symbol, you can supply a matrix of arbitrary character strings (in LaTeX notation), and these will be used as the elements of the matrix.

The function prints the resulting code to the console (by default). When the result is assigned to variable, you can send it to the clipboard using write_clip. Perhaps most convenient of all, the function can be used used in a markdown chunk in a Rmd or qmd document, e.g,

```{r results = "asis"}
symbolicMatrix("\lambda", nrow=2, ncol=2,
               diag=TRUE,
               lhs = "\boldsymbol{\Lambda}")
```

This generates $$ \boldsymbol{\Lambda} = \begin{pmatrix} \lambda_{1} & 0 \\ 0 & \lambda_{2} \\ \end{pmatrix} $$

Usage

symbolicMatrix(
  symbol = "x",
  nrow = "n",
  ncol = "m",
  matrix = "pmatrix",
  diag = FALSE,
  comma = FALSE,
  exponent,
  transpose = FALSE,
  prefix = "",
  suffix = "",
  lhs,
  print = TRUE
)

Arguments

symbol

name for matrix elements, character string. For LaTeX symbols, the backslash must be doubled because it is an escape character in R. That is, you must use symbol = "\\beta" to get \(\beta\). Alternatively, this can be an R matrix object, containing LaTeX code for the elements. For a row or column vector, use matrix(..., nrow=1) or matrix(..., ncol=1)

nrow

Number of rows, a single character representing rows symbolically, or an integer, generating that many rows.

ncol

Number of columns, a single character representing columns symbolically, or an integer, generating that many columns.

matrix

Character string giving the LaTeX matrix environment used in \begin{}, \end{}. Typically one of:

"pmatrix"

uses parentheses: "(", ")"

"bmatrix"

uses square brackets: "[", "]"

"Bmatrix"

uses braces: "{", "}"

"vmatrix"

uses vertical bars: "|", "|"

"Vmatrix"

uses double vertical bars: "||", "||"

"matrix"

generates a plain matrix without delimeters

diag

logical; if TRUE, off-diagonal elements are all 0 (and nrow must == ncol)

comma

logical; if TRUE, commas are inserted between row and column subscripts, as in x_{1,1}.

exponent

if specified, e.g., "-1", or "1/2", the exponent is applied to the matrix

transpose

if TRUE, the transpose symbol "\top" is appended to the matrix; this may also be a character string, e.g., "T", "\prime", "\textsf{T}" are commonly used.

prefix

optional character string to be pre-pended to each matrix element, e.g, to wrap each element in a function like "\sqrt" (but add braces)

suffix

optional character string to be appended to each matrix element, e.g., for exponents on each element

lhs

character; an optional LaTeX expression, e.g, "\boldsymbol{\Lamda}", for left-hand side of an equation with the generated matrix on the right-hand side.

print

logical; print the LaTeX code for the matrix on the console?; default: TRUE

Value

Returns invisibly the LaTeX representation of the matrix as a character string. If you assign to a variable, you can use write_clip to copy it to the clipboard.

As a side-effect, by default (unless print = FALSE) the function uses cat() to display the result at the console.

Details

This implementation assumes that the LaTeX amsmath package will be available because it uses the shorthands \begin{pmatrix}, ... rather than

\left(
  \begin{array}(ccc)
  ...
  \end{array}
\right)

.

You may need to use extra_dependencies: ["amsmath"] in your YAML header of a Rmd or qmd file.

You can actually supply a numeric matrix as the symbol, but the result will not be pretty unless the elements are integers or are rounded. For a LaTeX representation of general numeric matrices, use matrix2latex.

See also

Author

John Fox

Examples

symbolicMatrix()
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & \cdots & x_{1m} \\ 
#>   x_{21} & x_{22} & \cdots & x_{2m} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   x_{n1} & x_{n2} & \cdots & x_{nm} \\ 
#> \end{pmatrix}

# return value
mat <- symbolicMatrix(print = FALSE)
str(mat)
#>  chr "\\begin{pmatrix} \n  x_{11} & x_{12} & \\cdots & x_{1m} \\\\ \n  x_{21} & x_{22} & \\cdots & x_{2m} \\\\ \n  \\"| __truncated__
cat(mat)
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & \cdots & x_{1m} \\ 
#>   x_{21} & x_{22} & \cdots & x_{2m} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   x_{n1} & x_{n2} & \cdots & x_{nm} \\ 
#> \end{pmatrix}
# copy to clipboard
#clipr::write_clip(mat)    # this can't be done in non-interactive mode

# can use a complex symbol
symbolicMatrix("\\widehat{\\beta}", 2, 4)
#> \begin{pmatrix} 
#>   \widehat{\beta}_{11} & \widehat{\beta}_{12} & \widehat{\beta}_{13} & \widehat{\beta}_{14} \\ 
#>   \widehat{\beta}_{21} & \widehat{\beta}_{22} & \widehat{\beta}_{23} & \widehat{\beta}_{24} \\ 
#> \end{pmatrix}

# numeric rows/cols
symbolicMatrix(ncol=3)
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & x_{13} \\ 
#>   x_{21} & x_{22} & x_{23} \\ 
#>   \vdots & \vdots & \vdots \\ 
#>   x_{n1} & x_{n2} & x_{n3} \\ 
#> \end{pmatrix}
symbolicMatrix(nrow=4)
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & \cdots & x_{1m} \\ 
#>   x_{21} & x_{22} & \cdots & x_{2m} \\ 
#>   x_{31} & x_{32} & \cdots & x_{3m} \\ 
#>   x_{41} & x_{42} & \cdots & x_{4m} \\ 
#> \end{pmatrix}
symbolicMatrix(nrow=4, ncol=4)
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & x_{13} & x_{14} \\ 
#>   x_{21} & x_{22} & x_{23} & x_{24} \\ 
#>   x_{31} & x_{32} & x_{33} & x_{34} \\ 
#>   x_{41} & x_{42} & x_{43} & x_{44} \\ 
#> \end{pmatrix}

# diagonal matrices
symbolicMatrix(nrow=3, ncol=3, diag=TRUE)
#> \begin{pmatrix} 
#>   x_{1} & 0     & 0     \\ 
#>   0     & x_{2} & 0     \\ 
#>   0     & 0     & x_{3} \\ 
#> \end{pmatrix}
symbolicMatrix(nrow="n", ncol="n", diag=TRUE)
#> \begin{pmatrix} 
#>   x_{1} & 0     & \cdots & 0     \\ 
#>   0     & x_{2} & \cdots & 0     \\ 
#>   \vdots & \vdots & \ddots & \vdots \\ 
#>   0     & 0     & \cdots & x_{n} \\ 
#> \end{pmatrix}

# commas, exponents, transpose
symbolicMatrix("\\beta", comma=TRUE, exponent="-1")
#> \begin{pmatrix} 
#>   \beta_{1,1} & \beta_{1,2} & \cdots & \beta_{1,m} \\ 
#>   \beta_{2,1} & \beta_{2,2} & \cdots & \beta_{2,m} \\ 
#>   \vdots      & \vdots      &        & \vdots      \\ 
#>   \beta_{n,1} & \beta_{n,2} & \cdots & \beta_{n,m} \\ 
#> \end{pmatrix}^{-1}
symbolicMatrix("\\beta", comma=TRUE, transpose=TRUE)
#> \begin{pmatrix} 
#>   \beta_{1,1} & \beta_{1,2} & \cdots & \beta_{1,m} \\ 
#>   \beta_{2,1} & \beta_{2,2} & \cdots & \beta_{2,m} \\ 
#>   \vdots      & \vdots      &        & \vdots      \\ 
#>   \beta_{n,1} & \beta_{n,2} & \cdots & \beta_{n,m} \\ 
#> \end{pmatrix}^\top
symbolicMatrix("\\beta", comma=TRUE, exponent="-1", transpose=TRUE)
#> \begin{pmatrix} 
#>   \beta_{1,1} & \beta_{1,2} & \cdots & \beta_{1,m} \\ 
#>   \beta_{2,1} & \beta_{2,2} & \cdots & \beta_{2,m} \\ 
#>   \vdots      & \vdots      &        & \vdots      \\ 
#>   \beta_{n,1} & \beta_{n,2} & \cdots & \beta_{n,m} \\ 
#> \end{pmatrix}^{{-1^\top}}

# for a row/column vector, wrap in matrix()
symbolicMatrix(matrix(LETTERS[1:4], nrow=1))
#> \begin{pmatrix} 
#> A & B & C & D \\ 
#> \end{pmatrix}
symbolicMatrix(matrix(LETTERS[1:4], ncol=1))
#> \begin{pmatrix} 
#> A \\ 
#> B \\ 
#> C \\ 
#> D \\ 
#> \end{pmatrix}

# represent the SVD, X = U D V'  symbolically
X <- symbolicMatrix("x", "n", "p")
#> \begin{pmatrix} 
#>   x_{11} & x_{12} & \cdots & x_{1p} \\ 
#>   x_{21} & x_{22} & \cdots & x_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   x_{n1} & x_{n2} & \cdots & x_{np} \\ 
#> \end{pmatrix}
U <- symbolicMatrix("u", "n", "k")
#> \begin{pmatrix} 
#>   u_{11} & u_{12} & \cdots & u_{1k} \\ 
#>   u_{21} & u_{22} & \cdots & u_{2k} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   u_{n1} & u_{n2} & \cdots & u_{nk} \\ 
#> \end{pmatrix}
D <- symbolicMatrix("\\lambda", "k", "k", diag=TRUE)
#> \begin{pmatrix} 
#>   \lambda_{1} & 0           & \cdots      & 0           \\ 
#>   0           & \lambda_{2} & \cdots      & 0           \\ 
#>   \vdots      & \vdots      & \ddots      & \vdots      \\ 
#>   0           & 0           & \cdots      & \lambda_{k} \\ 
#> \end{pmatrix}
V <- symbolicMatrix("v", "k", "p", transpose = TRUE)
#> \begin{pmatrix} 
#>   v_{11} & v_{12} & \cdots & v_{1p} \\ 
#>   v_{21} & v_{22} & \cdots & v_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   v_{k1} & v_{k2} & \cdots & v_{kp} \\ 
#> \end{pmatrix}^\top
cat("\\mathrm{SVD:}\n", X, "=\n", U, D, V)
#> \mathrm{SVD:}
#>  \begin{pmatrix} 
#>   x_{11} & x_{12} & \cdots & x_{1p} \\ 
#>   x_{21} & x_{22} & \cdots & x_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   x_{n1} & x_{n2} & \cdots & x_{np} \\ 
#> \end{pmatrix}
#>  =
#>  \begin{pmatrix} 
#>   u_{11} & u_{12} & \cdots & u_{1k} \\ 
#>   u_{21} & u_{22} & \cdots & u_{2k} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   u_{n1} & u_{n2} & \cdots & u_{nk} \\ 
#> \end{pmatrix}
#>  \begin{pmatrix} 
#>   \lambda_{1} & 0           & \cdots      & 0           \\ 
#>   0           & \lambda_{2} & \cdots      & 0           \\ 
#>   \vdots      & \vdots      & \ddots      & \vdots      \\ 
#>   0           & 0           & \cdots      & \lambda_{k} \\ 
#> \end{pmatrix}
#>  \begin{pmatrix} 
#>   v_{11} & v_{12} & \cdots & v_{1p} \\ 
#>   v_{21} & v_{22} & \cdots & v_{2p} \\ 
#>   \vdots & \vdots &        & \vdots \\ 
#>   v_{k1} & v_{k2} & \cdots & v_{kp} \\ 
#> \end{pmatrix}^\top

# specify left hand side
symbolicMatrix("\\lambda", 3, 3, diag=TRUE, lhs = "\\boldsymbol{\\Lambda}")
#> \boldsymbol{\Lambda} = 
#> \begin{pmatrix} 
#>   \lambda_{1} & 0           & 0           \\ 
#>   0           & \lambda_{2} & 0           \\ 
#>   0           & 0           & \lambda_{3} \\ 
#> \end{pmatrix}

# supply a matrix for 'symbol'
m <- matrix(c(
  "\\alpha", "\\beta",
  "\\gamma", "\\delta",
  "\\epsilon", "\\pi",
  0 , 0), 4, 2, byrow=TRUE)
symbolicMatrix(m)
#> \begin{pmatrix} 
#> \alpha   & \beta  \\ 
#> \gamma   & \delta \\ 
#> \epsilon & \pi    \\ 
#> 0        & 0      \\ 
#> \end{pmatrix}

# Identity matrix
symbolicMatrix(diag(3), lhs = "\\mathbf{I}_3")
#> \mathbf{I}_3 = 
#> \begin{pmatrix} 
#> 1 & 0 & 0 \\ 
#> 0 & 1 & 0 \\ 
#> 0 & 0 & 1 \\ 
#> \end{pmatrix}

# prefix / suffix
symbolicMatrix(prefix="\\sqrt{", suffix="}")
#> \begin{pmatrix} 
#>   \sqrt{x_{11}} & \sqrt{x_{12}} & \cdots & \sqrt{x_{1m}} \\ 
#>   \sqrt{x_{21}} & \sqrt{x_{22}} & \cdots & \sqrt{x_{2m}} \\ 
#>   \vdots        & \vdots        &        & \vdots        \\ 
#>   \sqrt{x_{n1}} & \sqrt{x_{n2}} & \cdots & \sqrt{x_{nm}} \\ 
#> \end{pmatrix}
symbolicMatrix(suffix="^{1/2}")
#> \begin{pmatrix} 
#>   x_{11}^{1/2} & x_{12}^{1/2} & \cdots & x_{1m}^{1/2} \\ 
#>   x_{21}^{1/2} & x_{22}^{1/2} & \cdots & x_{2m}^{1/2} \\ 
#>   \vdots       & \vdots       &        & \vdots       \\ 
#>   x_{n1}^{1/2} & x_{n2}^{1/2} & \cdots & x_{nm}^{1/2} \\ 
#> \end{pmatrix}